Struct libgir::analysis::object::Info

pub struct Info {

    pub base: InfoBase,
    pub c_type: String,
    pub c_class_type: Option<String>,
    pub get_type: String,
    pub is_interface: bool,
    pub is_fundamental: bool,
    pub supertypes: Vec<StatusedTypeId>,
    pub final_type: bool,
    pub generate_trait: bool,
    pub trait_name: String,
    pub has_constructors: bool,
    pub has_functions: bool,
    pub signals: Vec<Info>,
    pub notify_signals: Vec<Info>,
    pub properties: Vec<Property>,
    pub builder_properties: Vec<(Vec<Property>, TypeId)>,
    pub builder_postprocess: Option<String>,
    pub child_properties: ChildProperties,
    pub signatures: Signatures,
    pub ref_fn: Option<String>,
    pub unref_fn: Option<String>,

}

Fields
base: InfoBasec_type: Stringc_class_type: Option<String>get_type: Stringis_interface: boolis_fundamental: boolsupertypes: Vec<StatusedTypeId>final_type: boolgenerate_trait: booltrait_name: Stringhas_constructors: boolhas_functions: boolsignals: Vec<Info>notify_signals: Vec<Info>properties: Vec<Property>builder_properties: Vec<(Vec<Property>, TypeId)>builder_postprocess: Option<String>child_properties: ChildPropertiessignatures: Signaturesref_fn: Option<String>

Specific to fundamental types

unref_fn: Option<String>

Specific to fundamental types

Implementations

impl Info

pub fn has_signals(&self) -> bool

pub fn should_generate_impl_block(&self) -> bool

Whether we should generate an impl block for this object We don’t generate an impl block if the type doesn’t have any of the followings:

Constructors / Functions / Builder properties (no build function)

Is a final type & doesn’t have either methods / properties / child properties / signals

pub fn need_generate_inherent(&self) -> bool

pub fn need_generate_trait(&self) -> bool

pub fn has_action_signals(&self) -> bool

pub fn function_location(&self, fn_info: &Info) -> LocationInObject

Returns the location of the function within this object

pub fn generate_doc_link_info(
&self,
fn_info: &Info
) -> (Cow<'_, str>, Cow<'_, str>)

Generate doc name based on function location within this object See also Self::function_location(). Returns (item/crate path including type name, just the type name)