Struct libgir::Config

pub struct Config {

    pub work_mode: WorkMode,
    pub girs_dirs: Vec<PathBuf>,
    pub girs_version: Vec<GirVersion>,
    pub library_name: String,
    pub library_version: String,
    pub target_path: PathBuf,
    pub auto_path: PathBuf,
    pub doc_target_path: PathBuf,
    pub external_libraries: Vec<ExternalLibrary>,
    pub objects: BTreeMap<String, GObject>,
    pub min_cfg_version: Version,
    pub use_gi_docgen: bool,
    pub make_backup: bool,
    pub generate_safety_asserts: bool,
    pub deprecate_by_min_version: bool,
    pub show_statistics: bool,
    pub concurrency: Concurrency,
    pub single_version_file: Option<PathBuf>,
    pub trust_return_value_nullability: bool,
    pub disable_format: bool,
    pub split_build_rs: bool,
    pub extra_versions: Vec<Version>,
    pub lib_version_overrides: HashMap<Version, Version>,
    pub feature_dependencies: HashMap<Version, Vec<String>>,
    pub external_docs_url: Option<String>,

}

Fields
work_mode: WorkModegirs_dirs: Vec<PathBuf>girs_version: Vec<GirVersion>library_name: Stringlibrary_version: Stringtarget_path: PathBufauto_path: PathBuf

Path where files generated in normal and sys mode

doc_target_path: PathBufexternal_libraries: Vec<ExternalLibrary>objects: BTreeMap<String, GObject>min_cfg_version: Versionuse_gi_docgen: boolmake_backup: boolgenerate_safety_asserts: booldeprecate_by_min_version: boolshow_statistics: boolconcurrency: Concurrencysingle_version_file: Option<PathBuf>trust_return_value_nullability: booldisable_format: boolsplit_build_rs: boolextra_versions: Vec<Version>lib_version_overrides: HashMap<Version, Version>feature_dependencies: HashMap<Version, Vec<String>>external_docs_url: Option<String>

An url that will be inserted into the docs as link that links to another doc source, for example when builds on docs.rs are limited due to license issues.

Implementations

impl Config

pub fn new<'a, S, W>( config_file: S, work_mode: W, girs_dirs: &[String], library_name: S, library_version: S, target_path: S, doc_target_path: S, make_backup: bool, show_statistics: bool, disable_format: bool ) -> Result<Self, String>
where S: Into<Option<&'a str>>, W: Into<Option<WorkMode>>,

pub fn library_full_name(&self) -> String

pub fn filter_version(&self, version: Option<Version>) -> Option<Version>

pub fn find_ext_library( &self, namespace: &Namespace ) -> Option<&ExternalLibrary>

pub fn min_required_version( &self, env: &Env, ns_id: Option<NsId> ) -> Option<Version>

pub fn resolve_type_ids(&mut self, library: &Library)

pub fn check_disable_format(&mut self)