Struct libgir::analysis::functions::Info

pub struct Info {

    pub name: String,
    pub func_name: String,
    pub new_name: Option<String>,
    pub glib_name: String,
    pub status: GStatus,
    pub kind: FunctionKind,
    pub visibility: Visibility,
    pub type_name: Result,
    pub parameters: Parameters,
    pub ret: Info,
    pub bounds: Bounds,
    pub outs: Info,
    pub version: Option<Version>,
    pub deprecated_version: Option<Version>,
    pub not_version: Option<Version>,
    pub cfg_condition: Option<String>,
    pub assertion: SafetyAssertionMode,
    pub doc_hidden: bool,
    pub doc_trait_name: Option<String>,
    pub doc_struct_name: Option<String>,
    pub doc_ignore_parameters: HashSet<String>,
    pub async: bool,
    pub unsafe_: bool,
    pub trampoline: Option<AsyncTrampoline>,
    pub callbacks: Vec<Trampoline>,
    pub destroys: Vec<Trampoline>,
    pub remove_params: Vec<usize>,
    pub async_future: Option<AsyncFuture>,
    pub hidden: bool,
    pub commented: bool,
    pub ns_id: NsId,
    pub generate_doc: bool,

}

Fields
name: Stringfunc_name: Stringnew_name: Option<String>glib_name: Stringstatus: GStatuskind: FunctionKindvisibility: Visibilitytype_name: Resultparameters: Parametersret: Infobounds: Boundsouts: Infoversion: Option<Version>deprecated_version: Option<Version>not_version: Option<Version>cfg_condition: Option<String>assertion: SafetyAssertionModedoc_hidden: booldoc_trait_name: Option<String>doc_struct_name: Option<String>doc_ignore_parameters: HashSet<String>async: boolunsafe_: booltrampoline: Option<AsyncTrampoline>callbacks: Vec<Trampoline>destroys: Vec<Trampoline>remove_params: Vec<usize>async_future: Option<AsyncFuture>hidden: bool

Whether the function is hidden (an implementation detail) Like the ref/unref/copy/free functions

commented: bool

Whether the function can’t be generated

ns_id: NsId

In order to generate docs links we need to know in which namespace this potential global function is defined

generate_doc: bool
Implementations

impl Info

pub fn codegen_name(&self) -> &str

pub fn is_special(&self) -> bool

pub fn should_be_doc_linked(&self, env: &Env) -> bool

pub fn should_docs_be_generated(&self, env: &Env) -> bool

pub fn doc_link(
&self,
parent: Option<&str>,
visible_parent: Option<&str>,
is_self: bool
) -> String

pub fn is_async_finish(&self, env: &Env) -> bool